+++ src/Resources/contao/languages/de/tl_backendmenue_zuordnungen.php
@@ -3,9 +3,10 @@
 declare(strict_types=1);
 
 $GLOBALS['TL_LANG']['tl_backendmenue_zuordnungen']['module'] = ['Backend-Modul', 'Wähle ein Backend-Modul aus, das diesem Bereich zugeordnet werden soll'];
-$GLOBALS['TL_LANG']['tl_backendmenue_zuordnungen']['position'] = ['Position', 'Sortierposition des Moduls innerhalb des Bereichs'];
 $GLOBALS['TL_LANG']['tl_backendmenue_zuordnungen']['edit'] = ['Bearbeiten'];
 $GLOBALS['TL_LANG']['tl_backendmenue_zuordnungen']['delete'] = ['Löschen'];
 $GLOBALS['TL_LANG']['tl_backendmenue_zuordnungen']['show'] = ['Anzeigen'];
+$GLOBALS['TL_LANG']['tl_backendmenue_zuordnungen']['cut'] = ['Verschieben', 'Die Zuordnung ID %s verschieben'];
+$GLOBALS['TL_LANG']['tl_backendmenue_zuordnungen']['pasteafter'] = ['Danach einfügen', 'Nach der Zuordnung ID %s einfügen'];
 
 $GLOBALS['TL_LANG']['tl_backendmenue_zuordnungen']['general_legend'] = 'Allgemein';